New Room switching system BASE v1
Room switching is now switching scene and not location as the old system did.
The new room manager now adding scene to the Server than sends command to the Client to change the client scene. When client changes scene, he will only be seen the players that on the new scene. The new system now knows if a room already active, if room is active the Server won't create a new Scene.
When Client requests to change the scene the server will tell the client to change the scene in additive mode.
"RoomManager" component will manage the trigger function, The new scene the client will go to & the desired position the player will reposition to.

Changelog:
- New scene switching.
- Removed the old scene switching
- Scenes folder arrengement
- Input field are now clears after sending commands
-Chatlog button is now fits better when active for mods